pyproject.toml
[tool.poetry]
name = "django-erp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
dependency-injector = "^4.41.0"
requests = "^2.31.0"
pydantic = "^2.7.0"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"
marshmallow = "^3.21.1"
sqlalchemy-serializer = "^1.4.12"
jsonpickle = "^3.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
